pyproject.toml
[tool.poetry]
name = "cli-template"
version = "0.1.0"
description = "Template for all CLI based projects"
authors = ["CubeVic <victoraguirre.f@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = ">=3.11"
typer = ">=0.15.4,<0.16.0"
rich = ">=14.0.0,<15.0.0"
python-dotenv = "^1.1.0"
pydantic = "^2.11.4"
pydantic-settings = "^2.9.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
hypothesis = "^6.131.18"

[tool.poetry.scripts]
cli-template = "app.main:main"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
